refactor: change style naming for constants and move to the separate file
