+++ src/constants/links.ts
@@ -1,0 +1,22 @@
+const HEADER_NAVIGATION = [
+  { name: "Home", href: "#", current: true },
+  { name: "World", href: "#", current: false },
+  { name: "Real Estate", href: "#", current: false },
+  { name: "Finance", href: "#", current: false },
+];
+
+const FOOTER_NAVIGATION = [
+  { name: "About Us", href: "#" },
+  { name: "Help", href: "#" },
+  { name: "Advertise", href: "#" },
+  { name: "Privacy Policy", href: "#" },
+  { name: "Terms of Service", href: "#" },
+];
+
+const SOCIAL_LINKS = [
+  { alt: "Twitter", src: "/twitter.svg", href: "#" },
+  { alt: "Facebook", src: "/facebook.svg", href: "#" },
+  { alt: "Instagram", src: "/instagram.svg", href: "#" },
+];
+
+export { HEADER_NAVIGATION, FOOTER_NAVIGATION, SOCIAL_LINKS };
