+++ src/layouts/root-layout/RootLayout.tsx
@@ -1,13 +1,13 @@
-import { Footer } from "@/components/common/footer/Footer";
 import { Header } from "@/components/common/header/Header";
+import { Footer } from "@/components/common/footer/Footer";
 
-import { openSans400 } from "@/styles/fonts";
+import { OPEN_SANS_400 } from "@/styles/fonts";
 import { ReactFCWithChildren } from "./types";
 
 const RootLayout: ReactFCWithChildren = ({ children }) => {
   return (
     <div
-      className={`${openSans400.className} flex flex-col justify-between h-screen`}
+      className={`${OPEN_SANS_400.className} flex flex-col justify-between h-screen`}
     >
       <Header />
       <main className="grow">{children}</main>
