+++ src/styles/fonts.ts
@@ -1,13 +1,13 @@
 import { Open_Sans } from "next/font/google";
 
-const openSans400 = Open_Sans({
+const OPEN_SANS_400 = Open_Sans({
   subsets: ["latin"],
   weight: "400",
 });
 
-const openSans600 = Open_Sans({
+const OPEN_SANS_600 = Open_Sans({
   subsets: ["latin"],
   weight: "600",
 });
 
-export { openSans400, openSans600 };
+export { OPEN_SANS_400, OPEN_SANS_600 };
