+++ src/utils/filter-class-names.ts
@@ -1,0 +1,5 @@
+function filterClassNames(...classes: String[]) {
+  return classes.filter(Boolean).join(" ");
+}
+
+export { filterClassNames };
